+++ codegen/class.ts
@@ -84,7 +84,6 @@ export async function* generateClasses(
 ): AsyncIterable<string> {
   runtimePath = runtimePath.replace(/\/+$/, "");
   yield "// deno-lint-ignore-file ban-unused-ignore\n";
-  yield 'import { Temporal } from "@js-temporal/polyfill";\n';
   yield "// @ts-ignore TS7016\n";
   yield 'import jsonld from "jsonld";\n';
   yield `import { type LanguageTag, parseLanguageTag }